+++ .github/e2e-tests.yml
@@ -206,7 +206,7 @@ runner-test-matrix:
     test_go_project_path: integration-tests
 
   - id: system-tests/smoke/cre/por_test.go:TestCRE_OCR3_PoR_Workflow_SingleDon_MultipleWriters_MockedPrice
-    path: system-tests/tests/smoke/cre/por_multi_chain_test.go
+    path: system-tests/tests/smoke/cre/por_test.go
     test_env_type: docker
     runs_on: ubuntu-latest
     triggers:
@@ -221,7 +221,7 @@ runner-test-matrix:
       E2E_JD_VERSION: 0.9.0 # there is no latest tag for this repo, so we need to specify the version
       GITHUB_READ_TOKEN: "{{ env.GITHUB_API_TOKEN }}" # GATI-provided token that can read from capabilities and dev-platform repos
       CI: "true"
-      CTF_CONFIGS: "environment-multichain-one-don-ci.toml"
+      CTF_CONFIGS: "environment-one-don-multichain-ci.toml"
       # Anvil developer key, not a secret
       PRIVATE_KEY: "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
     test_go_project_path: system-tests
