+++ system-tests/tests/smoke/cre/cmd/configs/single-don.toml
@@ -44,13 +44,13 @@
       Level = 'debug'
       JSONConsole = true
 
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
+      [OCR2]
+      Enabled = true
+      DatabaseTimeout = '1s'
 
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
+      [P2P.V2]
+      Enabled = true
+      ListenAddresses = ['0.0.0.0:5001']
       """
 
   [[nodesets.node_specs]]
@@ -68,13 +68,13 @@
       Level = 'debug'
       JSONConsole = true
 
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
+      [OCR2]
+      Enabled = true
+      DatabaseTimeout = '1s'
 
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
+      [P2P.V2]
+      Enabled = true
+      ListenAddresses = ['0.0.0.0:5001']
       """
 
   [[nodesets.node_specs]]
@@ -92,13 +92,13 @@
       Level = 'debug'
       JSONConsole = true
 
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
+      [OCR2]
+      Enabled = true
+      DatabaseTimeout = '1s'
 
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
+      [P2P.V2]
+      Enabled = true
+      ListenAddresses = ['0.0.0.0:5001']
       """
 
   [[nodesets.node_specs]]
@@ -116,13 +116,13 @@
       Level = 'debug'
       JSONConsole = true
 
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
+      [OCR2]
+      Enabled = true
+      DatabaseTimeout = '1s'
 
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
+      [P2P.V2]
+      Enabled = true
+      ListenAddresses = ['0.0.0.0:5001']
       """
 
   [[nodesets.node_specs]]
@@ -140,11 +140,11 @@
       Level = 'debug'
       JSONConsole = true
 
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
+      [OCR2]
+      Enabled = true
+      DatabaseTimeout = '1s'
 
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
+      [P2P.V2]
+      Enabled = true
+      ListenAddresses = ['0.0.0.0:5001']
       """
