+++ system-tests/tests/smoke/cre/environment-capabilities-don-ci.toml
@@ -50,20 +50,9 @@
       # omitted on purpose, if not defined here it will be added by Go code if worker nodes from this DON have cron capability
       # capabilities = ['./cron']
       user_config_overrides = """
-      [Feature]
-			LogPoller = true
-
       [Log]
       Level = 'debug'
       JSONConsole = true
-
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
-
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
       """
 
 [[nodesets]]
@@ -82,20 +71,9 @@
     [nodesets.node_specs.node]
       image = "injected-at-runtime"
       user_config_overrides = """
-      [Feature]
-			LogPoller = true
-
       [Log]
       Level = 'debug'
       JSONConsole = true
-
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
-
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
       """
 
 [[nodesets]]
@@ -113,18 +91,7 @@
     [nodesets.node_specs.node]
       image = "injected-at-runtime"
       user_config_overrides = """
-      [Feature]
-			LogPoller = true
-
       [Log]
       Level = 'debug'
       JSONConsole = true
-
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
-
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
       """
