+++ system-tests/tests/smoke/cre/environment-gateway-don-ci.toml
@@ -17,7 +17,7 @@
   type = "docker"
 
 [[workflow_configs]]
-	workflow_name = "abcdefgasd"
+  workflow_name = "abcdefgasd"
   # 0x01 8e16c39e 0000 20 0000000000000000 = ETH/USD Benchmark Price with 0 decimals
   feed_id = "018e16c39e000320000000000000000000000000000000000000000000000000"
