+++ system-tests/tests/smoke/cre/environment-one-don-multichain-ci.toml
@@ -67,18 +67,7 @@
     [nodesets.node_specs.node]
       image = "injected-at-runtime"
       user_config_overrides = """
-      [Feature]
-			LogPoller = true
-
       [Log]
       Level = 'debug'
       JSONConsole = true
-
-			[OCR2]
-			Enabled = true
-			DatabaseTimeout = '1s'
-
-			[P2P.V2]
-			Enabled = true
-			ListenAddresses = ['0.0.0.0:5001']
       """
